+++ changelog.md
@@ -1,3 +1,9 @@
+## 9-September-2024
+
+#### Added
+1. Queue can be passed for the logger. It is Useful if the caller is listening to log queue.
+
+
 ## 5-April-2024
 
 #### Added